Backend: poprawki po review schematu
Team::matches() udawało relację — zwracało Buildera pod nazwą, która w with()
i withCount() rzuca LogicException. Rozbite na homeMatches()/awayMatches() jako
prawdziwe HasMany plus matchesQuery() dla guarda, z wyjaśnieniem, czemu jedno
HasMany po dwóch kluczach nie istnieje.

Reszta: kickoff_at jako datetime zamiast timestamp (TIMESTAMP kończy się w 2038
i podlega niejawnej konwersji stref, a ustaliliśmy UTC w bazie), numer zawodnika
poza zakres tinyinta, usunięte indeksy na user_id i match_id dublujące te spod
foreignId()->constrained(), usunięte stany factory bez konsumenta.

Dwa testy obiecywały w nazwie więcej, niż sprawdzały — poprawione nazwy i treść.
Doszły dwa brakujące: guard na turnieju (przypadek sztandarowy) i kod 409.

ERD w PLAN.md nie wymieniał Group.tournament_id, choć migracja go zakłada.

+++ backend/database/migrations/2026_08_30_100500_create_players_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('team_id')->constrained()->cascadeOnDelete();
             $table->string('name', 120);
-            $table->unsignedTinyInteger('number')->nullable();
+            $table->unsignedSmallInteger('number')->nullable();
             $table->string('position', 40)->nullable();
             $table->softDeletes();
             $table->timestamps();
